Move AOT stackmaps to a top-level construct.
Other compiler backends will (I think...) need access to this for their
deopt, so move this out of jitc_llvm into a top-level module `aotsmp.rs`.
