+++ ykrt/src/compile/jitc_llvm/mod.rs
@@ -28,8 +28,7 @@ use yksmp::LiveVar;
 use yksmp::StackMapParser;
 
 mod deopt;
-mod frame;
-pub(crate) use frame::load_aot_stackmaps;
+pub(crate) mod frame;
 mod llvmbridge;
 
 pub static LLVM_BITCODE: LazyLock<&[u8]> = LazyLock::new(|| {
